Firt rel. of certificate Issuer via DigiCert AP
This scripts allows an admin user to use the DigiCert APIs to manage request, authorization and download of a Digicert certificate for a given server.  Requires in Input 3 parameters:   FQDN of the server,  Admin User API key (token) and ORG-ID.  Credits to Marco Malavolti and Barbara Monticini for specific API calls examples.